backend/requirements.txt
# Core Dependencies
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.2
Werkzeug>=2.0.1
pytesseract>=0.3.10
pillow>=10.0.0
pdf2image>=1.16.0
requests==2.28.2
google-generativeai>=0.3.2

# Optional: For better OCR performance
opencv-python==4.8.1.78
numpy<2.0.0

# For development
gunicorn>=20.1.0

# Document Processing Support
python-docx>=0.8.11
docx2txt>=0.8
PyPDF2>=3.0.0
striprtf>=0.0.26

# Excel and Spreadsheet Support
openpyxl>=3.1.0
xlrd==1.2.0
pandas>=2.0.0

# PowerPoint Support
python-pptx>=0.6.21

# OpenDocument Support
odfpy>=1.4.1

# Web Scraping / HTML Parsing
beautifulsoup4==4.8.2
lxml>=4.9.0
textract>=1.6.5

# Text Processing
chardet~=3.0.4

# Translation
translate>=3.6.1

# NLP Libraries
nltk>=3.8.1
spacy>=3.7.2
langdetect>=1.0.9
pycountry>=22.3.5

# Additional format support
easyocr>=1.7.0 # Alternative OCR engine

# Auto-translation to English support
googletrans==4.0.0-rc1

# Testing
pytest
pytest-flask
pytest-mock
requests 